Separate out shader reflection
